pyproject.toml
[project]
name = "fedml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch==2.3.1",
    "numpy",
    "torchtext>=0.18.0",
    "pandas>=2.2.3",
    "matplotlib>=3.10.3",
    "torchdata==0.9.0",
    "portalocker>=3.1.1",
    "pyyaml>=6.0.2",
]
